// src/types/pbxmngt/attribute-filter.ts
/*
* Copyright 2021 ALE International
*
* Permission is hereby granted, free of charge, to any person obtaining a copy of this 
* software and associated documentation files (the "Software"), to deal in the Software 
* without restriction, including without limitation the rights to use, copy, modify, merge, 
* publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons 
* to whom the Software is furnished to do so, subject to the following conditions:
* 
* The above copyright notice and this permission notice shall be included in all copies or 
* substantial portions of the Software.
* 
* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING 
* BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND 
* NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, 
* DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, 
* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
*/

/**
 * AttributeFilter represents the possible operation to apply to an attribute to build a filter.
 * @see {@link PbxManagement.getObjectInstances}
 */
export enum AttributeFilter {
        
    /**
     * The attribute is equal to the value.
     */
    Equals = 0,

    /**
     * The attribute starts with the value.
     */
    StartsWith = 1,

    /**
     * The attributes ends with the value.
     */
    EndsWith = 2,

    /**
     * The attribute is not equals to the value.
     */
    NotEquals = 3,

    /**
     * The attribute is greather than or equals to the value.
     */
    GreatherThanOrEquals = 4,

    /**
     * The attribute is Less than or equals to the value.
     */
    LessThanOrEquals = 5
}